move to equipment.ts
